+++ lib/rubocop/cop/mixin/if_node.rb
@@ -11,8 +11,7 @@ module RuboCop
 
       def modifier_if?(node)
         node.loc.respond_to?(:keyword) &&
-          %w(if unless).include?(node.loc.keyword.source) &&
-          node.loc.respond_to?(:end) && node.loc.end.nil?
+          %w(if unless).include?(node.loc.keyword.source) && node.modifier_form?
       end
 
       def elsif?(node)