+++ lib/rubocop/cop/style/nested_modifier.rb
@@ -51,7 +51,7 @@ module RuboCop
         def modifier_while_or_until?(node)
           node.loc.respond_to?(:keyword) &&
             %w(while until).include?(node.loc.keyword.source) &&
-            node.loc.respond_to?(:end) && node.loc.end.nil?
+            node.modifier_form?
         end
 
         def autocorrect(node)