+++ spec/rubocop/cop/style/infinite_loop_spec.rb
@@ -29,7 +29,18 @@ describe RuboCop::Cop::Style::InfiniteLoop do
   it 'accepts Kernel#loop' do
     inspect_source(cop,
                    'loop { break if something }')
+    expect(cop.offenses).to be_empty
+  end
 
+  it 'accepts modifier while' do
+    inspect_source(cop,
+                   'something while true')
+    expect(cop.offenses).to be_empty
+  end
+
+  it 'accepts modifier until' do
+    inspect_source(cop,
+                   'something until false')
     expect(cop.offenses).to be_empty
   end